pyproject.toml
[build-system]
requires = ["hatchling", "ninja"]
build-backend = "hatchling.build"

[project]
name = "robotwin"
version = "0.1.0"
description = "RoboTwin: Dual-Arm Robot Benchmark with Generative Digital Twins"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "RoboTwin Team" }
]
dependencies = [
    "torch>=2.9.1",
    "torchvision>=0.19.1",
    "transforms3d>=0.4.2",
    "sapien>=3.0.0b1",
    "scipy>=1.10.1",
    "mplib>=0.2.1",
    "gymnasium>=0.29.1",
    "trimesh>=4.4.3",
    "open3d>=0.18.0",
    "imageio>=2.34.2",
    "pydantic>=2.12.5",
    "zarr>=3.1.5",
    "openai>=2.14.0",
    "huggingface_hub>=0.25.0",
    "h5py>=3.12.1",
    "azure-ai-inference>=1.0.0b9",
    "pyglet<2",
    "wandb>=0.18.0",
    "moviepy>=2.0.0",
    "termcolor>=2.4.0",
    "pandas>=2.0.0",
    "tqdm>=4.0.0",
    "av>=11.0.0",
    "matplotlib>=3.8.0",
    "ffmpeg-python>=0.2.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "PyYAML>=6.0.1",
    "dill>=0.3.8",
    "einops>=0.8.0",
    "diffusers>=0.29.0",
    "opencv-python>=4.10.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "transformers>=4.30.0",
    "timm>=0.9.0",
    "ninja",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.40.0",
    "numpy<2.0.0",
    "lightning>=2.6.0",
    "torch-geometric>=2.7.0",
    "ip",
    "warp-lang",
    "rerun-sdk>=0.23.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/robotwin"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
ip = { git = "https://github.com/matdmiller/instant_policy.git" }

[dependency-groups]
dev = [
    "editables>=0.5",
]

